# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Ruth Health Products & Services API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a product (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # SERVICES
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders (admin or own orders)
      responses:
        "200":
          description: List of orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create a checkout session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderInput"
      responses:
        "200":
          description: Checkout session URL
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CheckoutSession"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get an order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /admin/zoom-link:
    get:
      operationId: getZoomLink
      tags: [dashboard]
      summary: Get the default Zoom link
      responses:
        "200":
          description: Zoom link
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoomSettings"
    patch:
      operationId: updateZoomLink
      tags: [dashboard]
      summary: Update the default Zoom link (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ZoomSettingsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ZoomSettings"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
      required:
        - success

    # AUTH
...
    Product:
      type: object
      required: [id, name, description, imageUrl, guestPrice, memberPrice, commissionPct, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        guestPrice:
          type: number
        memberPrice:
          type: number
        commissionPct:
          type: number
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, description, guestPrice, memberPrice, commissionPct]
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        guestPrice:
          type: number
        memberPrice:
          type: number
        commissionPct:
          type: number
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        imageUrl:
          type: ["string", "null"]
        guestPrice:
          type: number
        memberPrice:
          type: number
        commissionPct:
          type: number

    # SERVICES
...
    OrderInput:
      type: object
      required: [itemType, itemId, shippingAddress]
      properties:
        itemType:
          type: string
          enum: [product, service]
        itemId:
          type: integer
        promoCodeUsed:
          type: ["string", "null"]
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
    Order:
      type: object
      required: [id, itemType, itemId, itemAmount, shippingFee, totalAmount, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: ["integer", "null"]
        itemType:
          type: string
          enum: [product, service]
        itemId:
          type: integer
        itemName:
          type: ["string", "null"]
        promoCodeUsed:
          type: ["string", "null"]
        itemAmount:
          type: number
        shippingFee:
          type: number
        totalAmount:
          type: number
        shippingAddress:
          type: ["string", "null"]
        stripePaymentId:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, paid, failed, refunded]
        createdAt:
          type: string
          format: date-time
    ShippingAddress:
      type: object
      required: [country]
      properties:
        line1:
          type: string
        city:
          type: string
        province:
          type: string
        postalCode:
          type: string
        country:
          type: string
    CheckoutSession:
      type: object
      required: [checkoutUrl, orderId]
      properties:
        checkoutUrl:
          type: string
        orderId:
          type: integer

    # BOOKING SLOTS
...
    ZoomSettings:
      type: object
      required: [zoomLink]
      properties:
        zoomLink:
          type: ["string", "null"]
    ZoomSettingsUpdate:
      type: object
      required: [zoomLink]
      properties:
        zoomLink:
          type: string